pyproject.toml
[project]
name = "tabscm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
